Cast to size_t

+++ apps/openmw/mwclass/npc.cpp
@@ -94,7 +94,7 @@ namespace
         for (const ESM::Attribute& attribute : attributes)
         {
             const ESM::Race::MaleFemale& value
-                = race->mData.mAttributeValues[ESM::Attribute::refIdToIndex(attribute.mId)];
+                = race->mData.mAttributeValues[static_cast<size_t>(ESM::Attribute::refIdToIndex(attribute.mId))];
             creatureStats.setAttribute(attribute.mId, male ? value.mMale : value.mFemale);
         }
